Feat/synchronization service

+++ Packages/Sources/Models/Entities/UserEntity.swift
@@ -16,7 +16,7 @@ public class UserEntity: NSManagedObject {
     }
 
     @NSManaged public var email: String?
-    @NSManaged public var id: Int64
+    @NSManaged public var id: Int
     @NSManaged public var name: String?
     @NSManaged public var username: String?
     @NSManaged public var todos: NSSet?
